Refactor JS folder structure by adding custom JS files
