+++ app/javascript/application.js
@@ -2,10 +2,4 @@
 import "@hotwired/turbo-rails"
 import "controllers"
 
-var animation = bodymovin.loadAnimation({
-  container: document.getElementById('bm'),
-  renderer: 'svg',
-  loop: true,
-  autoplay: true,
-  path: 'https://lottie.host/29bb7105-ced3-4178-88a2-f799e0361401/vQ08u7baXF.json'
-})
+import "custom/main"
